add csv to script for hadoop

+++ sentiment_analysis/sentiment_analysis.py
@@ -1,9 +1,12 @@
 import os
 from dotenv import load_dotenv
 from pymongo import MongoClient
-import time
+
 from textblob import TextBlob
 from bson.objectid import ObjectId
+import csv
+import glob
+import subprocess
 
 load_dotenv()
 
@@ -26,8 +29,46 @@ def getSubjectivity(tweet):
 def getPolarity(tweet):
     return TextBlob(tweet).sentiment.polarity
 
-# get all data
+def write_csv_text_data(list_text_data_hadoop):
+    """write csv for hadoop"""
+    # create folder csv if nos exists
+    final_directory = os.path.join(os.getcwd(), 'csv')
+    if not os.path.exists(final_directory):
+        os.makedirs(final_directory)
+
+    # get the last csv for get file number data
+    last_csv = glob.glob(f"{final_directory}/*")
+    # get number
+    if last_csv == [] :
+        number_last_csv = 0
+    else :
+        number_last_csv = last_csv[-1].split("/")[-1].split(".")[0].split("data")[-1]
+
+    # path csv create
+    path = f'{final_directory}/data{int(number_last_csv) + 1}.csv'
+
+    # write data text in csv
+    if list_text_data_hadoop != []:
+        keys = list_text_data_hadoop[0].keys()
+        with open(path, 'w', newline='') as output_file:
+            dict_writer = csv.DictWriter(output_file, keys)
+            dict_writer.writeheader()
+            dict_writer.writerows(list_text_data_hadoop)
+    
+
+
+# print("Text")
+# for t in text_data_db.find({}):
+#         print(t)
+# print("no text")
+# for t in text_data_db.find({}):
+#     if 'text' not in t:
+#         print(t)
+
+# # get all data
 all_data_text = text_data_db.find({})
+# list send on csv send on hadoop
+list_text_data_hadoop = []
 
 for data in all_data_text :
     # dict insert sentiment in mongo
@@ -49,8 +90,19 @@ for data in all_data_text :
             dict_sentiment["symbol"] = data["symbol"]
             dict_sentiment['_id'] = ObjectId()
             sentiment_data_db.insert_one(dict_sentiment)
+    # list send on csv send on hadoop
+    list_text_data_hadoop.append(data)
+    
+
+    if 'text' in data:
+        # delete data in text_data_db
+        text_data_db.delete_one({'text': data["text"]})
+
+# write csv (text data) for hadoop
+write_csv_text_data(list_text_data_hadoop)
 
 
-    # delete data in text_data_db
-    text_data_db.delete_one({'text': data["text"]})
+
+print(sentiment_data_db.find().count())
+
